database-caching/requirements.txt
PyMySQL==0.9.3
redis==3.0
redis-py-cluster
realpython-reader